# api/openapi.yaml
openapi: 3.1.0
info:
  title: Neurun API
  version: 0.1.0
  description: Build Python deployments and execute immutable builds.
servers:
  - url: http://localhost:1267
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      security: []
      operationId: register
      description: >-
        Create an account and sign it in. The organization is optional: supply
        organization_name to start one you own, or invite_token to join one you
        were invited to, or neither and create or join later. An account may own
        one organization and join any number. Per-IP limiting belongs at the
        edge, not in the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string, format: password}
                organization_name: {type: string}
                invite_token: {type: string}
      responses:
        "201":
          description: >-
            Account created. The session cookie is set unless sign-in failed, in
            which case session is absent and the caller signs in normally.
            member and organization are absent when no organization was named,
            or when the organization step failed — the account still stands, and
            the caller starts or joins one afterwards.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: {$ref: "#/components/schemas/User"}
                  organization: {$ref: "#/components/schemas/Organization"}
                  member: {$ref: "#/components/schemas/Member"}
                  session: {$ref: "#/components/schemas/Session"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "410": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /v1/invites/lookup:
    get:
      security: []
      operationId: lookupInvite
      description: >-
        Name the organization a token would join, without spending it, so a
        sign-up page can show it before the account exists.
      parameters:
        - name: token
          in: query
          required: true
          schema: {type: string}
      responses:
        "200":
          description: The invitation is live.
          content:
            application/json:
              schema:
                type: object
                required: [organization, email, role]
                properties:
                  organization: {$ref: "#/components/schemas/Organization"}
                  email: {type: string, format: email}
                  role: {$ref: "#/components/schemas/Role"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "410": {$ref: "#/components/responses/Problem"}
  /v1/auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string}
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionEnvelope"}
        "401": {$ref: "#/components/responses/Problem"}
...
  /v1/auth/session:
    get:
      security: []
      operationId: getSession
      responses:
        "200":
          description: Current session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionEnvelope"}
        "401": {$ref: "#/components/responses/Problem"}
  /v1/organizations:
    post:
      operationId: createOrganization
      description: >-
        Start an organization. Refused with 409 if the account already owns one.
        The session cookie is re-issued so the new membership is live at once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        "201":
          description: Organization created and the session moved into it.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Organization"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
    get:
      operationId: listOrganizations
      description: Every organization the signed-in account belongs to.
      responses:
        "200":
          description: Organizations.
          content:
            application/json:
              schema:
                type: object
                required: [organizations]
                properties:
                  organizations:
                    type: array
                    items: {$ref: "#/components/schemas/Organization"}
  /v1/organization:
    get:
      operationId: getOrganization
      description: The organization this credential acts in.
      responses:
        "200":
          description: Organization.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Organization"}
    patch:
      operationId: updateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        "200":
          description: Organization renamed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Organization"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/members:
    get:
      operationId: listMembers
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Members of this organization.
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items: {$ref: "#/components/schemas/Member"}
  /v1/members/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: {type: string}
    patch:
      operationId: updateMember
      description: >-
        Change a member role. The owner cannot be demoted, so an organization
        always has somebody who can administer it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: {$ref: "#/components/schemas/Role"}
      responses:
        "200":
          description: Member updated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        "409": {$ref: "#/components/responses/Problem"}
    delete:
      operationId: removeMember
      description: Remove a member. The owner cannot be removed.
      responses:
        "204": {description: Member removed.}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/invites:
    get:
      operationId: listInvites
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Invitations issued by this organization.
          content:
            application/json:
              schema:
                type: object
                required: [invites]
                properties:
                  invites:
                    type: array
                    items: {$ref: "#/components/schemas/Invite"}
    post:
      operationId: createInvite
      description: >-
        Offer membership to an address. The token is returned exactly once;
        only its digest is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, role]
              properties:
                email: {type: string, format: email}
                role: {$ref: "#/components/schemas/Role"}
      responses:
        "201":
          description: Invitation created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CreatedInvite"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /v1/invites/accept:
    post:
      operationId: acceptInvite
      description: Join an organization with an invitation, as an existing account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: {type: string}
      responses:
        "200":
          description: Membership granted.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "410": {$ref: "#/components/responses/Problem"}
  /v1/invites/{invite_id}:
    parameters:
      - name: invite_id
        in: path
        required: true
        schema: {type: string}
    delete:
      operationId: revokeInvite
      responses:
        "204": {description: Invitation revoked.}
        "404": {$ref: "#/components/responses/Problem"}
  /v1/projects:
    get:
      operationId: listProjects
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Projects visible to the principal.
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items: {$ref: "#/components/schemas/Project"}
  /v1/projects/{project_id}:
    parameters:
      - name: project_id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getProject
      responses:
        "200":
          description: Project.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
        "404": {$ref: "#/components/responses/Problem"}
    patch:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Project"}
  /v1/apps:
    get:
      operationId: listApps
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: name
          in: query
          description: Return the app with this exact project-scoped name.
          schema: {type: string, minLength: 1, maxLength: 120}
      responses:
        "200":
          description: Apps in the authenticated project.
          content:
            application/json:
              schema:
                type: object
                required: [apps]
                properties:
                  apps:
                    type: array
                    items: {$ref: "#/components/schemas/App"}
        "422": {$ref: "#/components/responses/Problem"}
    post:
      operationId: createApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, repository]
              properties:
                name: {type: string}
                repository:
                  type: string
                  description: >-
                    owner/name on GitHub. Required: an app has no other source
                    of code, and the installation must already be able to read
                    it.
                production_ref:
                  type: string
                  description: >-
                    The ref whose pushes deploy. Empty follows the repository's
                    default branch.
      responses:
        "201":
          description: App created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/apps/{app_id}:
    parameters:
      - name: app_id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getApp
      responses:
        "200":
          description: App.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "404": {$ref: "#/components/responses/Problem"}
    patch:
      operationId: updateApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        "200":
          description: App updated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/apps/{app_id}/repository:
    parameters:
      - name: app_id
        in: path
        required: true
        schema: {type: string}
    put:
      operationId: connectRepository
      description: >-
        Point the app at a GitHub repository, or disconnect it by sending an
        empty repository. The installation must already be able to read it: the
        ref is resolved before anything is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [repository]
              properties:
                repository:
                  type: string
                  description: owner/name, or empty to disconnect.
                production_ref:
                  type: string
                  description: >-
                    The ref a push must touch to deploy. Empty follows the
                    repository's default branch.
      responses:
        "200":
          description: App connected.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/github/installation:
    get:
      operationId: getInstallation
      responses:
        "200":
          description: The organization's GitHub App installation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Installation"}
        "409": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
    post:
      operationId: recordInstallation
      description: >-
        Record the installation GitHub redirects back with after somebody
        installs the app. The account it is installed on is read from GitHub,
        since the redirect carries only the installation id. Re-installing
        replaces the previous record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [installation_id]
              properties:
                installation_id: {type: string}
      responses:
        "201":
          description: Installation recorded.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Installation"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
    delete:
      operationId: deleteInstallation
      responses:
        "204": {description: Installation forgotten.}
        "409": {$ref: "#/components/responses/Problem"}
  /v1/github/repositories:
    get:
      operationId: listRepositories
      description: The repositories this organization's installation can read.
      responses:
        "200":
          description: Repositories the installation grants.
          content:
            application/json:
              schema:
                type: object
                required: [repositories]
                properties:
                  repositories:
                    type: array
                    items: {$ref: "#/components/schemas/Repository"}
        "409": {$ref: "#/components/responses/Problem"}
        "503": {$ref: "#/components/responses/Problem"}
...
  /v1/api-keys/{api_key_id}:
    parameters:
      - name: api_key_id
        in: path
        required: true
        schema: {type: string}
    delete:
      operationId: revokeAPIKey
      responses:
        "200":
          description: Key revoked.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/APIKey"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: Request could not be completed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  schemas:
...
    SessionEnvelope:
      type: object
      required: [session]
      properties:
        session: {$ref: "#/components/schemas/Session"}
...
    Role:
      type: string
      enum: [admin, operator, viewer]
    User:
      type: object
      required: [id, email, disabled, created_at, updated_at]
      properties:
        id: {type: string}
        email: {type: string, format: email}
        disabled: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Organization:
      type: object
      required: [id, name, owner_user_id, created_at, updated_at]
      properties:
        id: {type: string}
        name: {type: string}
        owner_user_id: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Member:
      type: object
      required: [user_id, email, role, owner, created_at, updated_at]
      properties:
        user_id: {type: string}
        email: {type: string, format: email}
        role: {$ref: "#/components/schemas/Role"}
        owner: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...